requirements.txt
Django
djangorestframework
python-dotenv
psycopg2-binary
django-filter
drf-yasg
django-unixtimestampfield
mysqlclient